# ModulosDespliegue/metabolite-separation-api/api/openapi.yaml
openapi: 3.0.0
info:
  title: Metabolite Separation API
  version: 1.0.0
  description: API for predicting and evaluating metabolite separation configurations.
servers:
  - url: /
paths:
  #/predict, /evaluate, /recommendFamily, /families, /experiments, /feedback
  /predict:
    post:
      tags:
        - Prediction
      summary: Predict the best configuration
      description: Given a metabolite class, returns the optimal separation configuration.
      operationId: predictPOST
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/class_request'
      responses:
        '200':
          description: Prediction result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/predict_response'
        '400':
          description: Bad request
        '404':
          description: Class not found
      x-eov-operation-handler: controllers/DefaultController

  /evaluate:
    post:
      tags:
        - Evaluate
      summary: Evaluate a configuration
      description: Evaluates a chromatographic configuration for a metabolite class using the alpha score.
      operationId: evaluatePOST
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/evaluate_request'
      responses:
        '200':
          description: Evaluation completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/evaluate_response'
        '400':
          description: Bad request
        '404':
          description: Class not found
      x-eov-operation-handler: controllers/DefaultController

  /recommendFamily:
    post:
      tags:
        - RecommendFamily
      summary: Recommend class from configuration
      description: Recommends the most suitable metabolite class for a given configuration.
      operationId: recommendFamilyPOST
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/configuration'
      responses:
        '200':
          description: Class recommendation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/recommend_family_response'
        '400':
          description: Bad request
        '404':
          description: No suitable class found
      x-eov-operation-handler: controllers/DefaultController

  /families:
    get:
      tags:
        - Families
      summary: List all metabolite classes
      description: Retrieves a list of available metabolite classes.
      operationId: familiesGET
      responses:
        '200':
          description: List of classes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/families_response'
        '404':
          description: No classes found
      x-eov-operation-handler: controllers/DefaultController

  /experiments:
    post:
      tags:
        - Experiments
      summary: Submit experimental data
      description: Adds new experiment data to the system.
      operationId: experimentsPOST
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/experiment_request_list'
      responses:
        '200':
          description: Experiment data added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/experiment_response'
        '400':
          description: Bad request
      x-eov-operation-handler: controllers/DefaultController

  /feedback:
    post:
      tags:
        - Feedback
      summary: Submit feedback
      description: Receives feedback on a predicted configuration including a user-assigned score.
      operationId: feedbackPOST
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/feedback_request'
      responses:
        '200':
          description: Feedback submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/feedback_response'
        '400':
          description: Bad request
      x-eov-operation-handler: controllers/DefaultController

    get:
      tags:
        - Feedback
      summary: Retrieve all feedback
      description: Returns all feedback stored in the system, optionally filtered by class.
      operationId: feedbackGET
      parameters:
        - in: query
          name: classname
          schema:
            type: string
          required: false
          description: Filter feedback by metabolite class
      responses:
        '200':
          description: Feedback retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/feedback_response_list'
        '400':
          description: Invalid query parameter
        '500':
          description: Internal server error
      x-eov-operation-handler: controllers/DefaultController
components:
  schemas:
    class_request:
      type: object
      required:
        - class
      properties:
        class:
          type: string
          example: "Phenols"

    predict_response:
      type: object
      required:
        - configuration
        - class
        - score
      properties:
        class:
          type: string
          description: The metabolite class evaluated.
        configuration:
          $ref: '#/components/schemas/configuration'
        score:
          type: number
          description: Predicted score for the configuration.

    evaluate_request:
      type: object
      required:
        - configuration
        - class
      properties:
        class:
          type: string
        configuration:
          $ref: '#/components/schemas/configuration'

    evaluate_response:
      type: object
      required:
        - score
      properties:
        score:
          type: number
          description: Evaluation score (alpha) for the configuration.

    recommend_family_response:
      type: object
      required:
        - class
        - score
      properties:
        class:
          type: string
        score:
          type: number

    families_response:
      type: array
      items:
        type: string

    feedback_request:
      type: object
      required:
        - configuration
        - class
        - score
        - feedback
      properties:
        configuration:
          $ref: '#/components/schemas/configuration'
        class:
          type: string
        score:
          type: number
        feedback:
          type: integer
          minimum: 0
          maximum: 10

    feedback_response:
      type: object
      properties:
        message:
          type: string
          example: "Feedback stored"

    feedback_response_list:
      type: array
      items:
        $ref: '#/components/schemas/feedback_request'

    experiment_request_list:
      type: array
      items:
        $ref: '#/components/schemas/experiment_request'

    experiment_response:
      type: object
      properties:
        message:
          type: string
          example: "Data received"

    experiment_request:
      type: object
      properties:
        name_metabolito:
          type: string
        formula:
          type: string
        rt:
          type: number
        smiles_std:
          type: string
        inchi_std:
          type: string
        inchikey_std:
          type: string
        classyfire:
          $ref: '#/components/schemas/classyfire'
        comment:
          type: string
        alternative_parents:
          type: string
        column:
          $ref: '#/components/schemas/experiment_column'
        gradiente:
          $ref: '#/components/schemas/experiment_gradient'

    classyfire:
      type: object
      properties:
        kingdom:
          type: string
        superclass:
          type: string
        class:
          type: string
        subclass:
          type: string
        level5:
          type: string
        level6:
          type: string

    experiment_column:
      type: object
      properties:
        usp_code:
          type: string
        name:
          type: string
        length:
          type: number
        particle_size:
          type: number
        temperature:
          type: number
        flowrate:
          type: number
        t0:
          type: number

    experiment_gradient:
      type: object
      properties:
        eluyente1:
          type: string
        eluyente1_gradiente:
          type: array
          items:
            type: number
        eluyente2:
          type: string
        eluyente2_gradiente:
          type: array
          items:
            type: number
        t_gradiente:
          type: array
          items:
            type: number
        ph1:
          type: number
        ph2:
          type: number

    configuration:
      type: object
      required:
        - eluent_1
        - eluent_2
        - eluent_1_ph
        - eluent_2_ph
        - column
        - gradient
      properties:
        eluent_1:
          type: string
        eluent_2:
          type: string
        eluent_1_ph:
          type: number
        eluent_2_ph:
          type: number
        column:
          $ref: '#/components/schemas/column'
        gradient:
          $ref: '#/components/schemas/gradient'

    column:
      type: object
      required:
        - column_usp_code
        - column_length
        - particle_size
        - column_temperature
        - column_flowrate
        - dead_time
      properties:
        column_usp_code:
          type: string
        column_length:
          type: number
        particle_size:
          type: number
        column_temperature:
          type: number
        column_flowrate:
          type: number
        dead_time:
          type: number

    gradient:
      type: object
      required:
        - duration
        - x0
        - x1
        - x2
        - x3
        - x4
        - x5
        - x6
        - x7
        - x8
        - x9
        - x10
      properties:
        duration:
          type: number
        x0:
          type: number
        x1:
          type: number
        x2:
          type: number
        x3:
          type: number
        x4:
          type: number
        x5:
          type: number
        x6:
          type: number
        x7:
          type: number
        x8:
          type: number
        x9:
          type: number
        x10:
          type: number

tags:
  - name: Prediction
    description: Operations related to predicting the best chromatographic configuration for a given metabolite class.
  - name: Evaluate
    description: Operations for evaluating the performance of a chromatographic configuration using an alpha score.
  - name: Families
    description: Endpoints to retrieve the list of supported metabolite classes in the system.
  - name: Experiments
    description: Endpoints to submit new experimental data to improve or validate the model.
  - name: RecommendFamily
    description: Operations that recommend the most suitable metabolite class for a given configuration.
  - name: Feedback
    description: Endpoints to submit or retrieve user feedback on predicted configurations.